% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-functions.R
\name{base_reached}
\alias{base_reached}
\title{Which base does a player who obtains a hit reach?}
\usage{
base_reached(player)
}
\arguments{
\item{player}{an integer from 1 to 9 which indicates which player to retrieve statistics about from the PadresBatting.csv data file}
}
\value{
1, 2, 3 or 4 indicating which base a player made it to
}
\description{
Which base does a player who obtains a hit reach?
}
\examples{
base_reached(1)
base_reached(2)
}
