% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-functions.R
\name{game}
\alias{game}
\title{Simulates One Game for the Padres}
\usage{
game()
}
\value{
an integer, 0 if they lost the game and 1 if they won
}
\description{
Simulates One Game for the Padres
}
\examples{
game()
}
