% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-functions.R
\name{get_hit}
\alias{get_hit}
\title{Simulates if a player got a hit}
\usage{
get_hit(OBP)
}
\arguments{
\item{OBP}{double, the On-base Percentage of a player}
}
\value{
0 or 1, if a player gets on base a 1 is returned. if they get out a 0 is returned
}
\description{
Simulates if a player got a hit
}
\examples{
get_hit(.300)
}
