% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-functions.R
\name{half_inning}
\alias{half_inning}
\title{Simulates the batting order of the Padres for a half-inning}
\usage{
half_inning(starting_player)
}
\arguments{
\item{starting_player}{an integer, which player begins the Inning at bat}
}
\value{
a list with values, hits = total hits acheived by each player,
last_player = the player who hit right before the inning ended,
and runs = the total runs scored by all the players
}
\description{
Simulates the batting order of the Padres for a half-inning
}
\examples{
half_inning(1)
half_inning(9)
}
