% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-functions.R
\name{next_player}
\alias{next_player}
\title{Iterate to the Next Player in the Lineup}
\usage{
next_player(last_player)
}
\arguments{
\item{last_player}{an integer from 1 to 9 corresponding with placement in the starting lineup}
}
\value{
if last_player is 9, it returns 1 otherwise it returns last_player + 1
}
\description{
Iterate to the Next Player in the Lineup
}
\examples{
next_player(1)
next_player(9)
}
