% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-functions.R
\name{runs}
\alias{runs}
\title{Runs scored in an Inning}
\usage{
runs(number_hits)
}
\arguments{
\item{number_hits}{an integer, the sum of all the bases each player made it to during an Inning}
}
\value{
an integer, the calculated number of runs scored for the team during a half-Inning
}
\description{
Runs scored in an Inning
}
\examples{
runs(3)
runs(4)
}
