% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-functions.R
\name{season}
\alias{season}
\title{Simulates One Season (based on 2019 stats) for the Padres}
\usage{
season()
}
\value{
an integer, the number of games won
}
\description{
Simulates One Season (based on 2019 stats) for the Padres
}
\examples{
season()
}
